requirements.txt
aniso8601==9.0.1
attrs==23.1.0
black==23.9.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
connexion==2.14.2
flagger==3.1.0
flasgger==0.9.7.1
Flask==2.2.5
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-MySQL==1.5.2
Flask-RESTful==0.3.10
flask-restplus==0.13.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mistune==3.0.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
SQLAlchemy==2.0.21
swagger-ui-bundle==0.0.9
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.2.3
